Added frame control, and signals for events
The renaming of model into spritemodel. I did this because using
lowercase model regularly was creating major conflicts with the built in
qmodel.

I also added the complete frame management backend, complete with
signals that just need to be hooked up to our buttons over on the
right(will do after this commit).

These signals connect to slots in the mainwindow.cpp, such that when a
frame is added, deleted, or selected, appropriate calls are made.

There are two signals:
	SendFrameListChanged -> sends the new index(for highlighting in the frame viewer), along with the new current pixmap(for display/updating in the canvas)
	SendUpdateAnimation -> sends the new current pixMap(so is only send with add and delete frame), so that the animation preview can update accordingly.

This class does not have a get currently, though the signal for updating
the animation works similarly.

We can now move on to implementing onion skinning, the frame viewer, and
the animation preview. The next big roadblock will be with the grid
background/onion skinning because we have to add another canvas. I will
talk to Calvin about this tomorrow.