Hooked up class to UI and implemented it all properly
Has classes and vars and so on. Should work right now, though I dont
know why it doenst save the current state of the frame before moving on
but that is a future me problem. As of right now it should be good to go
for use any anything and everything. I ran into some pretty big issues
but I sorted them out so we are doing pretty good now.