Change frame selection to pointer use pointers
I changed the frame selection methods to use pointers instead of making direct copies or indirect copies in insert and delete. I made add initial frame method to add the current frame in the starting canvas instead of using the duplicate method. I wasn't 100% sure on what the duplicate method's intention was so I left it out. It is still in the project in comments.