Loading Fixed and implemented
I added and adjusted the helpers in the spritemodel class so I can the getPixmapListValues() and getPixmapListObjects(). This is because I want to get the actual objects when I am setting the list in a load, but just the values when I am saving. Right now, the canvas will just update with the first frame in the list for the project. There is a bug in frame manage I found. After loading a project, if you go to delete your current frame, ie the first frame in the sequence, you get an index out of bounds error.